refactor: extract reconciliation.ts from daemon/main.ts
Move buffer reconciliation logic (REPLAYABLE_EVENT_TYPES, reconciledSessions,
replayEvent, reconcileSession, startup loop) out of main() into a
createReconciler() factory in src/daemon/reconciliation.ts.
