+++ src/daemon/main.ts
@@ -16,7 +16,7 @@ import { TranscriptMiner, extractTurnsFromBuffer } from '../capture/transcript-m
 import { createPerProjectAdapter, extensionForMimeType, type TranscriptTurn } from '../symbionts/adapter.js';
 import { claudeCodeAdapter } from '../symbionts/claude-code.js';
 import { findPackageRoot } from '../utils/find-package-root.js';
-import { EventBuffer, listBufferSessionIds, cleanStaleBuffers } from '../capture/buffer.js';
+import { EventBuffer, cleanStaleBuffers } from '../capture/buffer.js';
 import { loadManifests } from '../symbionts/detect.js';
 import { isPlanWriteEvent, capturePlan, type PlanWatchConfig } from './plan-capture.js';
 import type { RegisteredSession } from './lifecycle.js';
@@ -145,6 +145,7 @@ import {
   handleSubagentStart, handleSubagentStop, handleStopFailure,
   handleTaskCompleted, handleCompact,
 } from './event-handlers.js';
+import { createReconciler } from './reconciliation.js';
 export {
   handleUserPrompt, handleToolUse, handleStopBatches, handleToolFailure,
   handleSubagentStart, handleSubagentStop, handleStopFailure,
@@ -163,9 +164,6 @@ import path from 'node:path';
 
 /** Default limit for listing agent runs in the API. */
 const AGENT_RUNS_DEFAULT_LIMIT = 50;
-
-/** Event types replayed during buffer reconciliation. */
-const REPLAYABLE_EVENT_TYPES: ReadonlySet<string> = new Set(['user_prompt', 'tool_use', 'tool_failure']);
 
 // ---------------------------------------------------------------------------
 // Stale daemon cleanup
@@ -389,144 +387,8 @@ export async function main(): Promise<void> {
   const bufferDir = path.join(vaultDir, 'buffer');
   const sessionBuffers = new Map<string, EventBuffer>();
 
-  // Clean up stale buffer files (>24h) on startup
-  const startupCleanedCount = cleanStaleBuffers(bufferDir, STALE_BUFFER_MAX_AGE_MS);
-  if (startupCleanedCount > 0) {
-    logger.info(LOG_KINDS.CAPTURE_BUFFER, 'Buffer cleanup complete', { stale_removed: startupCleanedCount });
-  }
-
-  // Track sessions already reconciled this daemon lifetime to avoid
-  // redundant file reads (startup scan + register + event can all fire).
-  const reconciledSessions = new Set<string>();
-
-  // Reconcile all remaining buffer files on startup — recover events from
-  // sessions that had activity while the daemon was down.
-  for (const sessionId of listBufferSessionIds(bufferDir)) {
-    try {
-      reconcileSession(sessionId);
-    } catch (err) {
-      logger.warn(LOG_KINDS.LIFECYCLE_RECONCILE, 'Startup reconciliation failed', { session_id: sessionId, error: String(err) });
-    }
-  }
-
-  /**
-   * Replay a single buffer event into the DB via the appropriate handler.
-   *
-   * Shared between reconcileSession (buffer replay) and the live /events
-   * route to eliminate dispatch duplication.
-   *
-   * @returns 'prompt' | 'activity' | null indicating what was created.
-   */
-  function replayEvent(sessionId: string, event: Record<string, unknown>): 'prompt' | 'activity' | null {
-    if (event.type === 'user_prompt') {
-      if (isSystemMessage(String(event.prompt ?? ''))) return null;
-      handleUserPrompt(sessionId, String(event.prompt ?? ''));
-      return 'prompt';
-    }
-    if (event.type === 'tool_use') {
-      handleToolUse(
-        sessionId,
-        String(event.tool_name ?? ''),
-        event.tool_input,
-        typeof event.output_preview === 'string' ? event.output_preview : undefined,
-      );
-      return 'activity';
-    }
-    if (event.type === 'tool_failure') {
-      handleToolFailure(
-        sessionId,
-        String(event.tool_name ?? ''),
-        event.tool_input,
-        typeof event.error === 'string' ? event.error : undefined,
-        !!event.is_interrupt,
-      );
-      return 'activity';
-    }
-    return null;
-  }
-
-  /**
-   * Reconcile buffer events against DB state for a session.
-   *
-   * The buffer is the authoritative event log. The DB (prompt_batches +
-   * activities) is a derived view. After a daemon restart, the DB may be
-   * missing events the daemon didn't process while it was down.
-   *
-   * Activities belong to batches — they're linked via the latest open batch
-   * at insertion time. So we can't reconcile them separately. Instead, we
-   * find where the DB diverges from the buffer (by prompt count) and replay
-   * the FULL event stream from that point: prompts open batches, tool events
-   * attach to the open batch — exactly the normal flow.
-   */
-  function reconcileSession(sessionId: string): void {
-    if (reconciledSessions.has(sessionId)) return;
-    reconciledSessions.add(sessionId);
-
-    // Read buffer file directly — avoid EventBuffer constructor which reads
-    // the file to compute a count we don't need.
-    const bufferPath = path.join(bufferDir, `${sessionId}.jsonl`);
-    if (!fs.existsSync(bufferPath)) return;
-    const content = fs.readFileSync(bufferPath, 'utf-8').trim();
-    if (!content) return;
-
-    // Buffer files outlive session rows — sessions may have been manually
-    // deleted or cleaned up by the session cleanup job. Skip reconciliation
-    // for sessions that no longer exist rather than resurrecting them.
-    if (!getSession(sessionId)) {
-      logger.debug(LOG_KINDS.LIFECYCLE_RECONCILE, 'Skipping reconciliation for deleted session', { session_id: sessionId });
-      return;
-    }
-
-    const allEvents: Array<Record<string, unknown>> = content.split('\n').map((line) => JSON.parse(line));
-
-    // Find the divergence point: how many real prompts does the DB have?
-    const existingBatchCount = listBatchesBySession(sessionId).length;
-
-    let promptsSeen = 0;
-    let replayStartIndex = -1;
-
-    for (let i = 0; i < allEvents.length; i++) {
-      const e = allEvents[i];
-      if (e.type === 'user_prompt' && !isSystemMessage(String(e.prompt ?? ''))) {
-        promptsSeen++;
-        if (promptsSeen === existingBatchCount + 1) {
-          replayStartIndex = i;
-          break;
-        }
-      }
-    }
-
-    if (replayStartIndex === -1) return;
-
-    // Replay full event stream from the divergence point
-    const eventsToReplay = allEvents.slice(replayStartIndex).filter(
-      (e) => REPLAYABLE_EVENT_TYPES.has(String(e.type)),
-    );
-
-    let promptsRecovered = 0;
-    let activitiesRecovered = 0;
-
-    for (const event of eventsToReplay) {
-      try {
-        const result = replayEvent(sessionId, event);
-        if (result === 'prompt') promptsRecovered++;
-        else if (result === 'activity') activitiesRecovered++;
-      } catch (err) {
-        logger.warn(LOG_KINDS.LIFECYCLE_RECONCILE, 'Reconciliation: failed to replay event', {
-          type: String(event.type),
-          error: String(err),
-        });
-      }
-    }
-
-    if (promptsRecovered > 0 || activitiesRecovered > 0) {
-      logger.info(LOG_KINDS.LIFECYCLE_RECONCILE, 'Buffer reconciliation complete', {
-        session_id: sessionId,
-        prompts_recovered: promptsRecovered,
-        activities_recovered: activitiesRecovered,
-      });
-    }
-  }
+  const reconciler = createReconciler({ bufferDir, logger });
+  reconciler.runStartupReconciliation();
 
   // Route body schemas
   const RegisterBody = z.object({
@@ -570,7 +432,7 @@ export async function main(): Promise<void> {
     updateSession(session_id, { ended_at: null, status: 'active' });
 
     // Reconcile buffer against DB — recover prompts lost if daemon was down mid-session.
-    reconcileSession(session_id);
+    reconciler.reconcileSession(session_id);
 
     logger.info(LOG_KINDS.LIFECYCLE_REGISTER, 'Session registered', { session_id, branch, started_at: started_at ?? null });
 
@@ -599,7 +461,7 @@ export async function main(): Promise<void> {
     // Prune in-memory state
     sessionBuffers.delete(session_id);
     sessionTitleCache.delete(session_id);
-    reconciledSessions.delete(session_id);
+    reconciler.clearSession(session_id);
     server.updateDaemonJsonSessions(registry.sessions);
     logger.info(LOG_KINDS.LIFECYCLE_UNREGISTER, 'Session unregistered', { session_id });
 
@@ -640,7 +502,7 @@ export async function main(): Promise<void> {
       });
 
       // Reconcile buffer against DB — recover any prompts lost during downtime.
-      reconcileSession(event.session_id);
+      reconciler.reconcileSession(event.session_id);
     }
 
     // Persist to disk so events survive daemon restarts
